+++ packages/website/.gitignore
@@ -1,2 +1,3 @@
 .next
 node_modules
+public
