+++ packages/website/components/BlogSection.tsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useState } from "react";
-import { getPosts } from "./getPosts";
 
 interface Content {
   body: string;
@@ -75,81 +74,16 @@ function getDateTime(timestamp: string): string {
     .padStart(2, "0")}-${date.getDate().toString().padStart(2, "0")}`;
 }
 
-function checkIfPostAreSet(posts) {
-  if (posts.length > 0) {
-    return posts.map((post: Post) => (
-      <div
-        key={post.content.title}
-        className="flex flex-col overflow-hidden rounded-lg shadow-lg"
-      >
-        <div className="flex-shrink-0">
-          <a
-            href={"https://mirror.xyz/labs.taiko.eth/" + post.OriginalDigest}
-            target="_blank"
-          >
-            <img
-              className="w-full h-40 sm:h-64 lg:h-36 xl:h-48 object-cover object-top"
-              src={`https://ipfs.io/ipfs/${post.wnft.imageURI}`}
-              alt=""
-            />
-          </a>
-        </div>
-        <div className="flex flex-1 flex-col justify-between bg-white p-6 dark:bg-neutral-800">
-          <div className="flex-1">
-            <a
-              href={"https://mirror.xyz/labs.taiko.eth/" + post.OriginalDigest}
-              target="_blank"
-              className="mt-2 block"
-            >
-              <div className="text-xl font-semibold text-neutral-900 dark:text-neutral-200">
-                {post.content.title}
-              </div>
-              <div className="mt-3 text-base text-neutral-500 dark:text-neutral-300">
-                {post.wnft.description}
-              </div>
-            </a>
-          </div>
-          <div className="mt-6 flex items-center">
-            <div className="ml-3">
-              <div className="flex space-x-1 text-sm text-neutral-500 dark:text-neutral-400">
-                <time dateTime={getDateTime(`${post.content.timestamp}`)}>
-                  {getDate(`${post.content.timestamp}`)}
-                </time>
-                <span aria-hidden="true">&middot;</span>
-                <span>{getReadingTime(post.content.body) + " min read"}</span>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    ));
-  } else {
-    const amountOfPostsToDisplay = 3;
-    return [...Array(amountOfPostsToDisplay)].map((v, i) => (
-      <div
-        key={i}
-        className="flex flex-col overflow-hidden rounded-lg shadow-lg"
-      >
-        <div className="h-80 animate-pulse">
-          <div className="h-full w-full bg-neutral-800"></div>
-        </div>
-      </div>
-    ));
-  }
-}
-
 export default function BlogSection(): JSX.Element {
   const [posts, setPosts] = useState<Object[]>([]);
-
   useEffect(() => {
-    getPosts.then((result : Post[]) => {
-      // only use the last three
-      result = result.sort((a, b) => b.content.timestamp - a.content.timestamp);
-      result = result.slice(0, 3);
-      
-      setPosts(result);
-    });
-    // getting the information of the post via the arweave GraphQL and SDK
+    fetch("/posts.json")
+      .then((response) => response.json())
+      .then((json) => {
+        json = json.sort((a, b) => b.content.timestamp - a.content.timestamp);
+        json = json.slice(0, 3);
+        setPosts(json);
+      });
   });
 
   return (
@@ -171,7 +105,58 @@ export default function BlogSection(): JSX.Element {
         </div>
 
         <div className="mx-auto mt-12 grid max-w-lg gap-5 lg:max-w-none lg:grid-cols-3">
-          {checkIfPostAreSet(posts)}
+          {posts.map((post: Post) => (
+            <div
+              key={post.content.title}
+              className="flex flex-col overflow-hidden rounded-lg shadow-lg"
+            >
+              <div className="flex-shrink-0">
+                <a
+                  href={
+                    "https://mirror.xyz/labs.taiko.eth/" + post.OriginalDigest
+                  }
+                  target="_blank"
+                >
+                  <img
+                    className="w-full h-40 sm:h-64 lg:h-36 xl:h-48 object-cover object-top"
+                    src={`https://ipfs.io/ipfs/${post.wnft.imageURI}`}
+                    alt=""
+                  />
+                </a>
+              </div>
+              <div className="flex flex-1 flex-col justify-between bg-white p-6 dark:bg-neutral-800">
+                <div className="flex-1">
+                  <a
+                    href={
+                      "https://mirror.xyz/labs.taiko.eth/" + post.OriginalDigest
+                    }
+                    target="_blank"
+                    className="mt-2 block"
+                  >
+                    <div className="text-xl font-semibold text-neutral-900 dark:text-neutral-200">
+                      {post.content.title}
+                    </div>
+                    <div className="mt-3 text-base text-neutral-500 dark:text-neutral-300">
+                      {post.wnft.description}
+                    </div>
+                  </a>
+                </div>
+                <div className="mt-6 flex items-center">
+                  <div className="ml-3">
+                    <div className="flex space-x-1 text-sm text-neutral-500 dark:text-neutral-400">
+                      <time dateTime={getDateTime(`${post.content.timestamp}`)}>
+                        {getDate(`${post.content.timestamp}`)}
+                      </time>
+                      <span aria-hidden="true">&middot;</span>
+                      <span>
+                        {getReadingTime(post.content.body) + " min read"}
+                      </span>
+                    </div>
+                  </div>
+                </div>
+              </div>
+            </div>
+          ))}
         </div>
       </div>
     </div>
